+++ lib/ui/BrowsePanel.dart
@@ -171,6 +171,17 @@ class _BrowseContentWidgetState extends State<BrowseContentWidget> implements No
     ]);
 
     _contentCodes = buildContentCodes();
+
+    // #4527 Auto-expand single item sections
+    if (_contentCodes != null) {
+      for (String code in _contentCodes!) {
+        List<String>? entryCodes = _BrowseSection.buildBrowseEntryCodes(sectionId: code);
+        if (entryCodes?.length == 1) {
+          _expandedCodes.add(code);
+        }
+      }
+    }
+
     super.initState();
   }
 
@@ -619,7 +630,7 @@ class _BrowseEntry extends StatelessWidget {
       case "dinings.dinings_open":           _onTapDiningsOpen(context); break;
       case "dinings.my_dining":              _onTapMyDinings(context); break;
 
-      case "directory.app_directory":        _onTapAppDirectory(context); break;
+      case "directory.user_directory":       _onTapUserDirectory(context); break;
 
       case "events.event_feed":              _onTapEventFeed(context); break;
       case "events.my_events":               _onTapMyEvents(context); break;
@@ -878,9 +889,9 @@ class _BrowseEntry extends StatelessWidget {
     )));
   }
 
-  void _onTapAppDirectory(BuildContext context) {
-    Analytics().logSelect(target: "Illinois App Directory");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) { return ProfileDirectoryAccountsPanel(DirectoryAccounts.appDirectory); } ));
+  void _onTapUserDirectory(BuildContext context) {
+    Analytics().logSelect(target: "User Directory");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) { return ProfileDirectoryAccountsPanel(DirectoryAccounts.userDirectory); } ));
   }
 
   void _onTapLaundry(BuildContext context) {