+++ lib/ui/messages/MessagesDirectoryPanel.dart
@@ -4,23 +4,23 @@ import 'package:illinois/ui/messages/MessagesConversationPanel.dart';
 import 'package:illinois/ui/messages/MessagesWidgets.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryAccountsPage.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryPage.dart';
+import 'package:illinois/ui/profile/ProfileDirectoryWidgets.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
 import 'package:illinois/utils/AppUtils.dart';
-import 'package:rokwire_plugin/model/auth2.directory.dart';
 import 'package:rokwire_plugin/model/social.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/social.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
-import 'package:rokwire_plugin/utils/utils.dart';
 
 class MessagesDirectoryPanel extends StatefulWidget {
   final List<Conversation> recentConversations;
   final int conversationPageSize;
   final bool? unread;
   final void Function()? onTapBanner;
-  MessagesDirectoryPanel({Key? key, required this.recentConversations, required this.conversationPageSize, this.unread, this.onTapBanner}) : super(key: key);
+  MessagesDirectoryPanel({Key? key, required this.recentConversations, required this.conversationPageSize, this.unread, this.onTapBanner}) :
+    super(key: key);
 
   _MessagesDirectoryPanelState createState() => _MessagesDirectoryPanelState();
 }
@@ -28,11 +28,10 @@ class MessagesDirectoryPanel extends StatefulWidget {
 class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with TickerProviderStateMixin {
   final GlobalKey<RecentConversationsPageState> _recentPageKey = GlobalKey();
   final GlobalKey<ProfileDirectoryAccountsPageState> _allUsersPageKey = GlobalKey();
-  final ScrollController _scrollController = ScrollController();
+  final ScrollController _recentScrollController = ScrollController();
+  final ScrollController _allUsersScrollController = ScrollController();
   late TabController _tabController;
   int _selectedTab = 0;
-  Map<String, bool> _selectedAccountIds = {};
-  Map<String, List<String>> _selectedConversationIds = {};
 
   final List<String> _tabNames = [
     Localization().getStringEx('panel.messages.directory.tab.recent.label', 'Recent'),
@@ -51,7 +50,8 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
   void dispose() {
     _tabController.removeListener(_onTabChanged);
     _tabController.dispose();
-    _scrollController.dispose();
+    _recentScrollController.dispose();
+    _allUsersScrollController.dispose();
 
     super.dispose();
   }
@@ -75,25 +75,25 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
           controller: _tabController,
           physics: const NeverScrollableScrollPhysics(),
           children: [
-            _buildContinueButtonOverlay(_recentContent),
-            _buildContinueButtonOverlay(_allUsersContent),
+            _buildContinueButtonOverlay(_recentContent, scrollController: _recentScrollController),
+            _buildContinueButtonOverlay(_allUsersContent, scrollController: _allUsersScrollController),
           ],
         ),
       ),
     ],);
   }
 
-  Widget _buildContinueButtonOverlay(Widget content) {
+  Widget _buildContinueButtonOverlay(Widget content, { ScrollController? scrollController }) {
     return RefreshIndicator(onRefresh: _onRefresh, child:
       Stack(
           children: [
-            SingleChildScrollView(controller: _scrollController, physics: AlwaysScrollableScrollPhysics(), child:
+            SingleChildScrollView(controller: scrollController, physics: AlwaysScrollableScrollPhysics(), child:
               Padding(
-                padding: EdgeInsets.only(left: 16, right: 16, top: 24, bottom: _isAccountSelected ? 64 : 24,),
+                padding: EdgeInsets.only(left: 16, right: 16, top: 24, bottom: _hasSelectedAccounts ? 64 : 24,),
                 child: content,
               )
             ),
-            if (_isAccountSelected)
+            if (_hasSelectedAccounts)
               Column(
                 children: [
                   Expanded(child: Container()),
@@ -115,35 +115,45 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
   }
 
   Widget get _recentContent =>
-      RecentConversationsPage(key: _recentPageKey, recentConversations: widget.recentConversations, conversationPageSize: widget.conversationPageSize, scrollController: _scrollController,
-          selectedConversationIds: _selectedConversationIds, onToggleConversationSelection: _onToggleConversationSelected);
+    RecentConversationsPage(
+      key: _recentPageKey,
+      recentConversations: widget.recentConversations,
+      conversationPageSize: widget.conversationPageSize,
+      scrollController: _recentScrollController,
+      onSelectedConversationChanged: _onSelectedConversationChanged,
+    );
 
   Widget get _allUsersContent =>
-      ProfileDirectoryAccountsPage(DirectoryAccounts.appDirectory, key: _allUsersPageKey, scrollController: _scrollController, selectedAccountIds: _selectedAccountIds, onToggleAccountSelection: _onToggleAccountSelected,);
+    ProfileDirectoryAccountsPage(DirectoryAccounts.userDirectory,
+      displayMode: DirectoryDisplayMode.select,
+      key: _allUsersPageKey,
+      scrollController: _allUsersScrollController,
+      onSelectedAccountsChanged: _onSelectedAccountsChanged,
+    );
 
   Future<void> _onTapCreateConversation() async {
     // do not need to check for existing conversation with selected members because Social BB handles it
-    List<String> memberIds = [];
-    if (_selectedTab == 0) {
-      // recent
-      _selectedConversationIds.forEach((conversationId, members) {
-        memberIds.addAll(members);
-      });
-    } else {
-      // all users
-      _selectedAccountIds.forEach((accountId, selected) {
-        if (selected) {
-          memberIds.add(accountId);
+    Set<String>? memberIds = (_selectedTab == 0) ?
+      _recentPageKey.currentState?.selectedAccountIds :
+      _allUsersPageKey.currentState?.selectedAccountIds;
+
+    if ((memberIds != null) && memberIds.isNotEmpty) {
+      Conversation? conversation = await Social().createConversation(memberIds: memberIds.toList());
+      if (mounted) {
+        if (conversation != null) {
+          if (_selectedTab == 0) {
+            _allUsersPageKey.currentState?.clearSelectedIds();
+          }
+          else {
+            _recentPageKey.currentState?.clearSelectedIds();
+          }
+          Navigator.push(context, CupertinoPageRoute(builder: (context) => MessagesConversationPanel(conversation: conversation,)));
+        } else {
+          AppAlert.showDialogResult(context, Localization().getStringEx('panel.messages.directory.button.continue.failed.msg', 'Failed to create a conversation with the selected members.'));
         }
-      });
+      }
     }
-    Conversation? conversation = await Social().createConversation(memberIds: memberIds);
-    if (conversation != null) {
-      _selectedAccountIds.clear();
-      Navigator.push(context, CupertinoPageRoute(builder: (context) => MessagesConversationPanel(conversation: conversation,)));
-    } else {
-      AppAlert.showDialogResult(context, Localization().getStringEx('panel.messages.directory.button.continue.failed.msg', 'Failed to create a conversation with the selected members.'));
-    }
+
   }
 
   void _onTabChanged({bool manual = true}) {
@@ -154,25 +164,21 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     }
   }
 
-  void _onToggleAccountSelected(bool value, Auth2PublicAccount account) {
-    setState(() {
-      if (StringUtils.isNotEmpty(account.id)) {
-        _selectedAccountIds[account.id!] = value;
-      }
-    });
+  void _onSelectedAccountsChanged() {
+    setStateIfMounted(() {});
   }
 
-  void _onToggleConversationSelected(bool value, Conversation conversation) {
-    setState(() {
-      if (StringUtils.isNotEmpty(conversation.id)) {
-        _selectedConversationIds[conversation.id!] = conversation.memberIds ?? [];
-      }
-    });
+  void _onSelectedConversationChanged() {
+    setStateIfMounted(() {});
   }
 
-  Future<void> _onRefresh() async => _selectedTab == 0 ? _recentPageKey.currentState?.refresh() : _allUsersPageKey.currentState?.refresh();
+  Future<void> _onRefresh() async => (_selectedTab == 0) ?
+    _recentPageKey.currentState?.refresh() : _allUsersPageKey.currentState?.refresh();
 
-  bool get _isAccountSelected => _selectedAccountIds.containsValue(true);
+  bool get _hasSelectedAccounts => (_selectedTab == 0) ?
+    (_recentPageKey.currentState?.selectedConversationIds.isNotEmpty == true) :
+    (_allUsersPageKey.currentState?.selectedAccountIds.isNotEmpty == true);
+
 }
 
 enum MessagesDirectoryContentType { recent, all }